Fix some type conversion warnings

+++ revzip/src/revzip.cpp
@@ -550,7 +550,7 @@ void revZipExtractItemToVariable(char *p_arguments[], int p_argument_count, char
 	char *t_data = NULL;
 	if (t_result == NULL)
 	{
-		t_data = (char *)malloc(t_stat . size);
+		t_data = (char *)malloc((size_t)t_stat . size);
 		if (t_data == NULL)
 		{
 			t_result = strdup("ziperr,out of memory");
